# be-payment/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/payment/send": {
            "post": {
                "description": "Send Payment, with body request idUser Integer, idAccountFrom Integer, idAccountTo Integer, amount Integer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BE-PAYMENT"
                ],
                "summary": "Send Payment Endpoint",
                "parameters": [
                    {
                        "description": "idUser",
                        "name": "idUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "idAccountFrom",
                        "name": "idAccountFrom",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "idAccountTo",
                        "name": "idAccountTo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "amount",
                        "name": "amount",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/payment/withdraw": {
            "post": {
                "description": "Withdraw Payment, with body request idUser Integer, idAccountFrom Integer, amount Integer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BE-PAYMENT"
                ],
                "summary": "Withdraw Payment Endpoint",
                "parameters": [
                    {
                        "description": "idUser",
                        "name": "idUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "idAccountFrom",
                        "name": "idAccountFrom",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "amount",
                        "name": "amount",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}